Prod-44: Pull out logic from controller methods, clean up code
