+++ src/main/java/com/indiraactive/fulfillmentplatform/controller/InventoryUpdaterController.java
@@ -4,7 +4,8 @@ import com.indiraactive.fulfillmentplatform.dao.scriptRunAuditEntry.ScriptRunAud
 import com.indiraactive.fulfillmentplatform.dao.scriptRunAuditEntry.ScriptRunAuditEntryRepository;
 import com.indiraactive.fulfillmentplatform.dao.supplier.Supplier;
 import com.indiraactive.fulfillmentplatform.dao.supplier.SupplierRepository;
-import com.indiraactive.fulfillmentplatform.service.InventoryUpdater;
+import com.indiraactive.fulfillmentplatform.service.inventoryUpdater.InventoryUpdater;
+import com.indiraactive.fulfillmentplatform.service.inventoryUpdater.InventoryUpdaterManagerImpl;
 import com.indiraactive.fulfillmentplatform.viewModel.RunHistoryViewModel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
@@ -15,19 +16,7 @@ import java.util.List;
 
 public class InventoryUpdaterController {
     @Autowired
-    private ScriptRunAuditEntryRepository scriptRunAuditEntryRepository;
-
-    /**
-     * Runs the sync_inventory script based on a supplier provided to it
-     */
-    @Autowired
-    private InventoryUpdater inventoryUpdater;
-
-    /**
-     * Abstraction to physical database that creates, reads, and deletes suppliers
-     */
-    @Autowired
-    private SupplierRepository supplierRepository;
+    private InventoryUpdaterManagerImpl inventoryUpdaterManager;
 
     /**
      * Lets a user choose a supplier and update inventory based on what that selected supplier has
@@ -37,7 +26,7 @@ public class InventoryUpdaterController {
     @GetMapping("/inventoryUpdater")
     public String inventoryUpdater(Model model) {
         model.addAttribute("supplier", new Supplier());
-        model.addAttribute("suppliers", supplierRepository.findAll());
+        model.addAttribute("suppliers", inventoryUpdaterManager.getSuppliersToUpdate());
         return "inventoryUpdater";
     }
 
@@ -48,40 +37,16 @@ public class InventoryUpdaterController {
      */
     @PostMapping("/updateInventory")
     public String updateInventory(@ModelAttribute Supplier supplier) {
-        try {
-            System.out.println("STARTED RUNNING SYNC_INVENTORY.PY");
-            inventoryUpdater.updateInventory(supplier.getSupplierId());
-            System.out.println("FINISHED RUNNING SYNC_INVENTORY.PY");
-        } catch (Exception e) {
-            e.printStackTrace();
-            return "error";
+        if (inventoryUpdaterManager.updateInventory(supplier)) {
+            return "success";
         }
-        return "success";
+        return "error";
     }
 
     @RequestMapping(value="runHistory", method= RequestMethod.GET)
     public String runHistory(Model model, @RequestParam(value = "order", required = false) String order,
                              @RequestParam(value = "property", required = false) String property) {
-        List<Supplier> supplierRepositories = new LinkedList<>();
-        supplierRepository.findAll().forEach(supplierRepositories::add);
-
-        List<ScriptRunAuditEntry> scriptRunAuditEntries = new LinkedList<>();
-        if (order != null) {
-            if (order.equalsIgnoreCase("asc") && property.equalsIgnoreCase("startDate")) {
-                scriptRunAuditEntries.addAll(scriptRunAuditEntryRepository.findAllByOrderByStartDateTimeAsc());
-            } else if (order.equalsIgnoreCase("desc") && property.equalsIgnoreCase("startDate")) {
-                scriptRunAuditEntries.addAll(scriptRunAuditEntryRepository.findAllByOrderByStartDateTimeDesc());
-            } else if (order.equalsIgnoreCase("asc") && property.equalsIgnoreCase("endDate")) {
-                scriptRunAuditEntries.addAll(scriptRunAuditEntryRepository.findAllByOrderByFinishDateTimeAsc());
-            } else if (order.equalsIgnoreCase("desc") && property.equalsIgnoreCase("endDate")) {
-                scriptRunAuditEntries.addAll(scriptRunAuditEntryRepository.findAllByOrderByFinishDateTimeDesc());
-            }
-        } else {
-            scriptRunAuditEntries.addAll(scriptRunAuditEntryRepository.findAll());
-        }
-
-        RunHistoryViewModel runHistoryViewModel = new RunHistoryViewModel(scriptRunAuditEntries, supplierRepositories);
-        model.addAttribute("runHistoryViewModels", runHistoryViewModel.getRunHistoryModel());
+        model.addAttribute("runHistoryViewModels", inventoryUpdaterManager.getRunHistory(order, property));
 
         return "runHistory";
     }
